pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuck-vacancysoft-refined"
version = "0.1.0"
description = "Coverage-first job scraping pipeline redesign"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8,<3",
  "sqlalchemy>=2.0,<3",
  "alembic>=1.13,<2",
  "typer>=0.12,<1",
  "httpx>=0.27,<1",
  "playwright>=1.52,<2",
  "structlog>=24.1,<25",
  "rapidfuzz>=3.9,<4",
  "python-dateutil>=2.9,<3",
  "pyyaml>=6.0,<7",
  "tenacity>=9.0,<10",
  "openpyxl>=3.1,<4",
  "rich>=13.0,<15",
  "openai>=1.0,<2",
  "psycopg2-binary>=2.9,<3",
  "arq>=0.26,<1",
  "redis>=5.0,<6",
  "python-dotenv>=1.0,<2",
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.23,<1",
  "ruff>=0.6,<1",
  "mypy>=1.11,<2",
]

[project.scripts]
prospero = "vacancysoft.cli.app:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
# Repo root ("." ) is on the path so tests/test_config.py can import
# `configs.config` — the `configs/` directory lives at the repo root, not
# under src/. Without "." here, the console-script `pytest` invocation
# (used by CI) fails to collect test_config.py with ModuleNotFoundError.
# `python -m pytest` works either way because it adds CWD automatically.
pythonpath = [".", "src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
